Make IPTS work on 5.3
Since in 5.3 the support for GuC submission was (temporarily) removed from i915,
this takes the latest version of i915 from 5.2 adds it into the tree, and fixes
up all API differences. The result is a working IPTS touch module on kernel 5.3.

Note: The driver is in a `i915_legacy` directory, but the kernel module itself
is still named i915. This was done because it is not possible to generate a
patch that entirely removes i915 if it already has patches applied (either
through linux-stable updates or distribution kernels). Putting it into a new
directory solves that problem.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>