patch-chromeos: re-add "surface_sam: hps: use result of pci_enable_device()"
When building with CONFIG_ENABLE_MUST_CHECK=y and CONFIG_ERROR_ON_WARNING=y
(which are the default values on stock chromeos kernel config), build
fails.

I before built kernel with "CONFIG_ENABLE_MUST_CHECK is not set" and I
thought this was fixed but it turned out it's not true.

So, re-add this patch.

Signed-off-by: Tsuchiya Yuto (kitakar5525) <kitakar@gmail.com>